requirements.txt
dialogflow==0.7.2
google-api-core==1.16.0
google-auth==1.11.2
python-dotenv==0.12.0
python-telegram-bot==11.1.0
vk-api==11.7.0
